add config mapper dto

+++ src/main/java/com/info/markets/model/MarketResponseApiData.java
@@ -9,17 +9,15 @@ public class MarketResponseApiData {
     public Date date;
     public String symbol;
     public String exchange;
-    public int open;
-    public int high;
-    public int low;
-    public int close;
-    public int volume;
+    public long open;
+    public long high;
+    public long low;
+    public long close;
+    public long volume;
     public Double adj_high;
     public Double adj_close;
     public Double adj_open;
     public Double adj_low;
     public Double adj_volume;
     public Double split_factor;
-
-
 }
